Remove unsafety of mutating visitor
On a second look, Rust is right in this case and it is unsafe to pass same lifetime 🤦🏻‍♂️

The reason is that the owner might store a mutable reference it was given outside of the visitor callback (which was the whole point of using same lifetime in the first place), and now we'll have two mutable references to the same location - one outside and one inside the visitor - which is not allowed.

If we decide that we still need API that allows to store mutable references outside the visitor in the future, we can provide a separate method with signature similar to `(&'a mut T) -> Option<&'a mut T>` so that the owner can either indicate that it stored the mutable reference or return its ownership back to the visitor walker.

For now, however, let's just remove lifetime and make mutable references temporary.