Add handling for deeply nested blocks
For now implementing WasmbinEncode, WasmbinDecode, WasmbinVisit and Drop in a safe manner using heap-allocated stacks of blocks.

In the future will need to be extended to other types as well.

Also changed depth test to a benchmark that not only ensures we can handle ~infinite nesting, but also measures performance of handling such modules.