Add SIMD support
Unlike all other opcodecs, SIMD uses u32 (encoded as LEB128) for its list of operations. Because of this, the derive macro had to be extended to be generic over repr(ty) instead of allowing only repr(u8), and UnrecognizedDiscriminant had to be extended to accept i128 to cover any potential discriminants up to u64/i64.

Additionally, SIMD heavily uses byte arrays, so in this commit I'm adding Encode/Decode/Visit implementation for them for common lengths.