Breaking: make Vec<u8> length-prefixed
This swaps around semantics of length-prefixed blobs of bytes and unprefixed.

Since unprefixed bytes should be less common in the AST, it doesn't make sense to assign them to `Vec<u8>`, while the more common length-prefixed vectors need to be accessed via special `RawBlob` type.

This commit removes the `RawBlob` type, makes `Vec<u8>` length-prefixed for consistency with other vectors and strings, and introduces `UnparsedBytes` for unprefixed blobs of bytes which are only used for unparsed "leftovers" in couple of places.

This is pretty significant change, but worth it in the long run.